27. ReactJS + Material UI • Tổ chức cấu trúc dự án chuyên sâu (Board - Columns - Cards)
