+++ src/pages/Boards/BoardContent/ListColumns/Column/Column.jsx
@@ -1,0 +1,156 @@
+import DragHandleIcon from '@mui/icons-material/DragHandle'
+import Box from '@mui/material/Box'
+import Typography from '@mui/material/Typography'
+import AddCardIcon from '@mui/icons-material/AddCard'
+import Tooltip from '@mui/material/Tooltip'
+import Menu from '@mui/material/Menu'
+import MenuItem from '@mui/material/MenuItem'
+import Divider from '@mui/material/Divider'
+import Button from '@mui/material/Button'
+import ListItemText from '@mui/material/ListItemText'
+import ListItemIcon from '@mui/material/ListItemIcon'
+import ContentCut from '@mui/icons-material/ContentCut'
+import ContentCopy from '@mui/icons-material/ContentCopy'
+import ContentPaste from '@mui/icons-material/ContentPaste'
+import Cloud from '@mui/icons-material/Cloud'
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
+import { useState } from 'react'
+import DeleteForeverIcon from '@mui/icons-material/DeleteForever'
+import ListCards from './ListCards/ListCards'
+
+function Column() {
+  const [anchorEl, setAnchorEl] = useState(null)
+
+  const open = Boolean(anchorEl)
+  const handleClick = (event) => setAnchorEl(event.currentTarget)
+  const handleClose = () => setAnchorEl(null)
+
+  return (
+    <Box
+      sx={{
+        minWidth: '300px',
+        maxWidth: '300px',
+        bgcolor: (theme) => (theme.palette.mode === 'dark' ? '#333643' : '#ebecf0'),
+        ml: 2,
+        borderRadius: '6px',
+        height: 'fit-content',
+        maxHeight: (theme) => `calc(${theme.trello.boardContentHeight} - ${theme.spacing(5)})`
+      }}
+    >
+      {/* Box Column Header */}
+      <Box
+        sx={{
+          height: (theme) => theme.trello.columnHeaderHeight,
+          p: 2,
+          display: 'flex',
+          alignIems: 'center',
+          justifyContent: 'space-between'
+        }}
+      >
+        <Typography
+          variant='h6'
+          sx={{
+            fontSize: '1rem',
+            fontWeight: 'bold',
+            cursor: 'pointer'
+          }}
+        >
+          Column Title
+        </Typography>
+
+        <Box>
+          <Tooltip title='More options'>
+            <ExpandMoreIcon
+              sx={{ color: 'text.primary', cursor: 'pointer' }}
+              id='basic-column-dropdown'
+              aria-controls={open ? 'basic-menu-column-dropdown' : undefined}
+              aria-haspopup='true'
+              aria-expanded={open ? 'true' : undefined}
+              onClick={handleClick}
+            />
+          </Tooltip>
+          <Menu
+            id='basic-menu-column-dropdown'
+            anchorEl={anchorEl}
+            open={open}
+            onClose={handleClose}
+            MenuListProps={{
+              'aria-labelledby': 'basic-column-dropdown'
+            }}
+          >
+            <MenuItem>
+              <ListItemIcon>
+                <AddCardIcon fontSize='small' />
+              </ListItemIcon>
+              <ListItemText>Add new card</ListItemText>
+              <Typography variant='body2' color='text.secondary'>
+                ⌘X
+              </Typography>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <ContentCut fontSize='small' />
+              </ListItemIcon>
+              <ListItemText>Cut</ListItemText>
+              <Typography variant='body2' color='text.secondary'>
+                ⌘X
+              </Typography>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <ContentCopy fontSize='small' />
+              </ListItemIcon>
+              <ListItemText>Copy</ListItemText>
+              <Typography variant='body2' color='text.secondary'>
+                ⌘C
+              </Typography>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <ContentPaste fontSize='small' />
+              </ListItemIcon>
+              <ListItemText>Paste</ListItemText>
+              <Typography variant='body2' color='text.secondary'>
+                ⌘V
+              </Typography>
+            </MenuItem>
+            <Divider />
+            <MenuItem>
+              <ListItemIcon>
+                <DeleteForeverIcon fontSize='small' />
+              </ListItemIcon>
+              <ListItemText>Remove Archive this column</ListItemText>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <Cloud fontSize='small' />
+              </ListItemIcon>
+              <ListItemText>Archive this column</ListItemText>
+            </MenuItem>
+          </Menu>
+        </Box>
+      </Box>
+
+      {/* List Cards */}
+      <ListCards />
+
+      {/* Box Column Footer */}
+      <Box
+        sx={{
+          height: (theme) => theme.trello.columnFooterHeight,
+          p: 2,
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'space-between'
+        }}
+      >
+        <Button startIcon={<AddCardIcon />}>Add new card</Button>
+        <Tooltip title='Drag to move'>
+          <DragHandleIcon sx={{ cursor: 'pointer' }} />
+        </Tooltip>
+      </Box>
+    </Box>
+  )
+}
+
+export default Column
