--- src/pages/Boards/BoardContent/index.jsx
@@ -1,539 +1,0 @@
-import CommentIcon from '@mui/icons-material/Comment'
-import AttachmentIcon from '@mui/icons-material/Attachment'
-import GroupIcon from '@mui/icons-material/Group'
-import CardActions from '@mui/material/CardActions'
-import CardMedia from '@mui/material/CardMedia'
-import Card from '@mui/material/Card'
-import CardContent from '@mui/material/CardContent'
-import DragHandleIcon from '@mui/icons-material/DragHandle'
-import Box from '@mui/material/Box'
-import Typography from '@mui/material/Typography'
-import AddCardIcon from '@mui/icons-material/AddCard'
-import Tooltip from '@mui/material/Tooltip'
-import Menu from '@mui/material/Menu'
-import MenuItem from '@mui/material/MenuItem'
-import Divider from '@mui/material/Divider'
-import Button from '@mui/material/Button'
-import ListItemText from '@mui/material/ListItemText'
-import ListItemIcon from '@mui/material/ListItemIcon'
-import ContentCut from '@mui/icons-material/ContentCut'
-import ContentCopy from '@mui/icons-material/ContentCopy'
-import ContentPaste from '@mui/icons-material/ContentPaste'
-import Cloud from '@mui/icons-material/Cloud'
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
-import { useState } from 'react'
-import DeleteForeverIcon from '@mui/icons-material/DeleteForever'
-
-const COLUMN_HEADER_HEIGHT = '50px'
-const COLUMN_FOOTER_HEIGHT = '56px'
-
-function BoardContent() {
-  const [anchorEl, setAnchorEl] = useState(null)
-
-  const open = Boolean(anchorEl)
-  const handleClick = (event) => setAnchorEl(event.currentTarget)
-  const handleClose = () => setAnchorEl(null)
-
-  return (
-    <Box
-      sx={{
-        bgcolor: (theme) => (theme.palette.mode === 'dark' ? '#34495e' : '#1976d2'),
-        width: '100%',
-        height: (theme) => theme.trello.boardContentHeight,
-        p: '10px 0'
-      }}
-    >
-      <Box
-        sx={{
-          bgcolor: 'inherit',
-          width: '100%',
-          height: '100%',
-          display: 'flex',
-          overflowX: 'auto',
-          overflowY: 'hidden',
-          '&::-webkit-scrollbar-track': { m: 2 }
-        }}
-      >
-        {/* Box Column 01 */}
-        <Box
-          sx={{
-            minWidth: '300px',
-            maxWidth: '300px',
-            bgcolor: (theme) => (theme.palette.mode === 'dark' ? '#333643' : '#ebecf0'),
-            ml: 2,
-            borderRadius: '6px',
-            height: 'fit-content',
-            maxHeight: (theme) => `calc(${theme.trello.boardContentHeight} - ${theme.spacing(5)})`
-          }}
-        >
-          {/* Box Column Header */}
-          <Box
-            sx={{
-              height: COLUMN_HEADER_HEIGHT,
-              p: 2,
-              display: 'flex',
-              alignIems: 'center',
-              justifyContent: 'space-between'
-            }}
-          >
-            <Typography
-              variant='h6'
-              sx={{
-                fontSize: '1rem',
-                fontWeight: 'bold',
-                cursor: 'pointer'
-              }}
-            >
-              Column Title
-            </Typography>
-
-            <Box>
-              <Tooltip title='More options'>
-                <ExpandMoreIcon
-                  sx={{ color: 'text.primary', cursor: 'pointer' }}
-                  id='basic-column-dropdown'
-                  aria-controls={open ? 'basic-menu-column-dropdown' : undefined}
-                  aria-haspopup='true'
-                  aria-expanded={open ? 'true' : undefined}
-                  onClick={handleClick}
-                />
-              </Tooltip>
-              <Menu
-                id='basic-menu-column-dropdown'
-                anchorEl={anchorEl}
-                open={open}
-                onClose={handleClose}
-                MenuListProps={{
-                  'aria-labelledby': 'basic-column-dropdown'
-                }}
-              >
-                <MenuItem>
-                  <ListItemIcon>
-                    <AddCardIcon fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Add new card</ListItemText>
-                  <Typography variant='body2' color='text.secondary'>
-                    ⌘X
-                  </Typography>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentCut fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Cut</ListItemText>
-                  <Typography variant='body2' color='text.secondary'>
-                    ⌘X
-                  </Typography>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentCopy fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Copy</ListItemText>
-                  <Typography variant='body2' color='text.secondary'>
-                    ⌘C
-                  </Typography>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentPaste fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Paste</ListItemText>
-                  <Typography variant='body2' color='text.secondary'>
-                    ⌘V
-                  </Typography>
-                </MenuItem>
-                <Divider />
-                <MenuItem>
-                  <ListItemIcon>
-                    <DeleteForeverIcon fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Remove Archive this column</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <Cloud fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Archive this column</ListItemText>
-                </MenuItem>
-              </Menu>
-            </Box>
-          </Box>
-
-          {/* Box List Card */}
-          <Box
-            sx={{
-              p: '0 5px', // padding
-              m: '0 5px', // margin
-              display: 'flex',
-              flexDirection: 'column',
-              gap: 1,
-              overflowX: 'hidden',
-              overflowY: 'auto',
-              maxHeight: (theme) => `calc(
-              ${theme.trello.boardContentHeight} -
-              ${theme.spacing(5)} -
-              ${COLUMN_HEADER_HEIGHT} -
-              ${COLUMN_FOOTER_HEIGHT}
-            )`,
-              '&::-webkit-scrollbar-thumb': { backgroundColor: '#ced0da' },
-              '&::-webkit-scrollbar-thumb:hover': { backgroundColor: '#dfc2cf' }
-            }}
-          >
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardMedia
-                sx={{ height: 140 }}
-                image='https://trungquandev.com/wp-content/uploads/2022/07/fair-mern-stack-advanced-banner-trungquandev-scaled.jpg'
-                title='green iguana'
-              />
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>TrungQuanDev MERN Stack</Typography>
-              </CardContent>
-              <CardActions sx={{ p: '0 4px 8px 4px' }}>
-                <Button size='small' startIcon={<GroupIcon />}>
-                  20
-                </Button>
-                <Button size='small' startIcon={<CommentIcon />}>
-                  15
-                </Button>
-                <Button size='small' startIcon={<AttachmentIcon />}>
-                  10
-                </Button>
-              </CardActions>
-            </Card>
-
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-          </Box>
-
-          {/* Box Column Footer */}
-          <Box
-            sx={{
-              height: COLUMN_FOOTER_HEIGHT,
-              p: 2,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'space-between'
-            }}
-          >
-            <Button startIcon={<AddCardIcon />}>Add new card</Button>
-            <Tooltip title='Drag to move'>
-              <DragHandleIcon sx={{ cursor: 'pointer' }} />
-            </Tooltip>
-          </Box>
-        </Box>
-
-        {/* Box Column 02 */}
-        <Box
-          sx={{
-            minWidth: '300px',
-            maxWidth: '300px',
-            bgcolor: (theme) => (theme.palette.mode === 'dark' ? '#333643' : '#ebecf0'),
-            ml: 2,
-            borderRadius: '6px',
-            height: 'fit-content',
-            maxHeight: (theme) => `calc(${theme.trello.boardContentHeight} - ${theme.spacing(5)})`
-          }}
-        >
-          {/* Box Column Header */}
-          <Box
-            sx={{
-              height: COLUMN_HEADER_HEIGHT,
-              p: 2,
-              display: 'flex',
-              alignIems: 'center',
-              justifyContent: 'space-between'
-            }}
-          >
-            <Typography
-              variant='h6'
-              sx={{
-                fontSize: '1rem',
-                fontWeight: 'bold',
-                cursor: 'pointer'
-              }}
-            >
-              Column Title
-            </Typography>
-
-            <Box>
-              <Tooltip title='More options'>
-                <ExpandMoreIcon
-                  sx={{ color: 'text.primary', cursor: 'pointer' }}
-                  id='basic-column-dropdown'
-                  aria-controls={open ? 'basic-menu-column-dropdown' : undefined}
-                  aria-haspopup='true'
-                  aria-expanded={open ? 'true' : undefined}
-                  onClick={handleClick}
-                />
-              </Tooltip>
-              <Menu
-                id='basic-menu-column-dropdown'
-                anchorEl={anchorEl}
-                open={open}
-                onClose={handleClose}
-                MenuListProps={{
-                  'aria-labelledby': 'basic-column-dropdown'
-                }}
-              >
-                <MenuItem>
-                  <ListItemIcon>
-                    <AddCardIcon fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Add new card</ListItemText>
-                  <Typography variant='body2' color='text.secondary'>
-                    ⌘X
-                  </Typography>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentCut fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Cut</ListItemText>
-                  <Typography variant='body2' color='text.secondary'>
-                    ⌘X
-                  </Typography>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentCopy fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Copy</ListItemText>
-                  <Typography variant='body2' color='text.secondary'>
-                    ⌘C
-                  </Typography>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentPaste fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Paste</ListItemText>
-                  <Typography variant='body2' color='text.secondary'>
-                    ⌘V
-                  </Typography>
-                </MenuItem>
-                <Divider />
-                <MenuItem>
-                  <ListItemIcon>
-                    <DeleteForeverIcon fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Remove Archive this column</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <Cloud fontSize='small' />
-                  </ListItemIcon>
-                  <ListItemText>Archive this column</ListItemText>
-                </MenuItem>
-              </Menu>
-            </Box>
-          </Box>
-
-          {/* Box List Card */}
-          <Box
-            sx={{
-              p: '0 5px', // padding
-              m: '0 5px', // margin
-              display: 'flex',
-              flexDirection: 'column',
-              gap: 1,
-              overflowX: 'hidden',
-              overflowY: 'auto',
-              maxHeight: (theme) => `calc(
-              ${theme.trello.boardContentHeight} -
-              ${theme.spacing(5)} -
-              ${COLUMN_HEADER_HEIGHT} -
-              ${COLUMN_FOOTER_HEIGHT}
-            )`,
-              '&::-webkit-scrollbar-thumb': { backgroundColor: '#ced0da' },
-              '&::-webkit-scrollbar-thumb:hover': { backgroundColor: '#dfc2cf' }
-            }}
-          >
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardMedia
-                sx={{ height: 140 }}
-                image='https://trungquandev.com/wp-content/uploads/2022/07/fair-mern-stack-advanced-banner-trungquandev-scaled.jpg'
-                title='green iguana'
-              />
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>TrungQuanDev MERN Stack</Typography>
-              </CardContent>
-              <CardActions sx={{ p: '0 4px 8px 4px' }}>
-                <Button size='small' startIcon={<GroupIcon />}>
-                  20
-                </Button>
-                <Button size='small' startIcon={<CommentIcon />}>
-                  15
-                </Button>
-                <Button size='small' startIcon={<AttachmentIcon />}>
-                  10
-                </Button>
-              </CardActions>
-            </Card>
-
-            <Card
-              sx={{
-                cursor: 'pointer',
-                boxShadow: '0 1px 1px rgba(0,0,0,0.2)',
-                overflow: 'unset'
-              }}
-            >
-              <CardContent sx={{ p: 1.5, '&:last-child': { p: 1.5 } }}>
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-          </Box>
-
-          {/* Box Column Footer */}
-          <Box
-            sx={{
-              height: COLUMN_FOOTER_HEIGHT,
-              p: 2,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'space-between'
-            }}
-          >
-            <Button startIcon={<AddCardIcon />}>Add new card</Button>
-            <Tooltip title='Drag to move'>
-              <DragHandleIcon sx={{ cursor: 'pointer' }} />
-            </Tooltip>
-          </Box>
-        </Box>
-      </Box>
-    </Box>
-  )
-}
-
-export default BoardContent
